完善离线包脚本

+++ offline/setup-offline-to-m2.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
+
 mkdir -p ~/.m2/offline/repository
+
+if [ -f "~/.m2/settings.xml" ];then
 cp -f ~/.m2/settings.xml ~/.m2/settings.xml.bak
+fi
+
 cp -f settings.xml ~/.m2/
 cp -rf repository ~/.m2/offline/
